Hide ARKG details from public previewSign API
